fix(test): 修复权限测试中的类型转换问题

+++ app/src/test/java/dev/woms/mumdroid/ChanACLTest.kt
@@ -11,73 +11,73 @@ class ChanACLTest {
     @Test
     fun kick_requiresKickBanOrWrite() {
         assertFalse(ChanACL.canKick(0))
-        assertFalse(ChanACL.canKick(ChanACL.MUTE_DEAFEN))
-        assertTrue(ChanACL.canKick(ChanACL.KICK))
-        assertTrue(ChanACL.canKick(ChanACL.BAN))
-        assertTrue(ChanACL.canKick(ChanACL.WRITE))
-        assertTrue(ChanACL.canKick(ChanACL.KICK or ChanACL.BAN or ChanACL.WRITE))
+        assertFalse(ChanACL.canKick(ChanACL.MUTE_DEAFEN.toLong()))
+        assertTrue(ChanACL.canKick(ChanACL.KICK.toLong()))
+        assertTrue(ChanACL.canKick(ChanACL.BAN.toLong()))
+        assertTrue(ChanACL.canKick(ChanACL.WRITE.toLong()))
+        assertTrue(ChanACL.canKick((ChanACL.KICK or ChanACL.BAN or ChanACL.WRITE).toLong()))
     }
 
     @Test
     fun ban_requiresBanOrWrite() {
         assertFalse(ChanACL.canBan(0))
-        assertFalse(ChanACL.canBan(ChanACL.KICK))
-        assertFalse(ChanACL.canBan(ChanACL.MUTE_DEAFEN))
-        assertTrue(ChanACL.canBan(ChanACL.BAN))
-        assertTrue(ChanACL.canBan(ChanACL.WRITE))
+        assertFalse(ChanACL.canBan(ChanACL.KICK.toLong()))
+        assertFalse(ChanACL.canBan(ChanACL.MUTE_DEAFEN.toLong()))
+        assertTrue(ChanACL.canBan(ChanACL.BAN.toLong()))
+        assertTrue(ChanACL.canBan(ChanACL.WRITE.toLong()))
     }
 
     @Test
     fun muteDeafen_isChannelScoped() {
-        assertFalse(ChanACL.canMuteDeafen(ChanACL.WRITE))
-        assertTrue(ChanACL.canMuteDeafen(ChanACL.MUTE_DEAFEN))
+        assertFalse(ChanACL.canMuteDeafen(ChanACL.WRITE.toLong()))
+        assertTrue(ChanACL.canMuteDeafen(ChanACL.MUTE_DEAFEN.toLong()))
         assertEquals(0x20, ChanACL.MOVE)
     }
 
     @Test
     fun move_requiresMoveBit() {
         assertFalse(ChanACL.canMove(0))
-        assertFalse(ChanACL.canMove(ChanACL.MUTE_DEAFEN))
-        assertTrue(ChanACL.canMove(ChanACL.MOVE))
-        assertTrue(ChanACL.canMove(ChanACL.WRITE))
-        assertTrue(ChanACL.canMove(ChanACL.WRITE or ChanACL.MOVE))
+        assertFalse(ChanACL.canMove(ChanACL.MUTE_DEAFEN.toLong()))
+        assertTrue(ChanACL.canMove(ChanACL.MOVE.toLong()))
+        assertTrue(ChanACL.canMove(ChanACL.WRITE.toLong()))
+        assertTrue(ChanACL.canMove((ChanACL.WRITE or ChanACL.MOVE).toLong()))
     }
 
     @Test
     fun prioritySpeaker_requiresWriteOrMuteDeafen() {
         assertFalse(ChanACL.canPrioritySpeaker(0))
-        assertFalse(ChanACL.canPrioritySpeaker(ChanACL.KICK))
-        assertTrue(ChanACL.canPrioritySpeaker(ChanACL.MUTE_DEAFEN))
-        assertTrue(ChanACL.canPrioritySpeaker(ChanACL.WRITE))
-        assertTrue(ChanACL.canPrioritySpeaker(ChanACL.WRITE or ChanACL.MUTE_DEAFEN))
+        assertFalse(ChanACL.canPrioritySpeaker(ChanACL.KICK.toLong()))
+        assertTrue(ChanACL.canPrioritySpeaker(ChanACL.MUTE_DEAFEN.toLong()))
+        assertTrue(ChanACL.canPrioritySpeaker(ChanACL.WRITE.toLong()))
+        assertTrue(ChanACL.canPrioritySpeaker((ChanACL.WRITE or ChanACL.MUTE_DEAFEN).toLong()))
     }
 
     @Test
     fun offerMute_selfOnlyWhenAlreadySilenced() {
         val md = ChanACL.MUTE_DEAFEN
         assertFalse(ChanACL.canOfferMute(0, isSelf = false, muted = false, suppressed = false))
-        assertTrue(ChanACL.canOfferMute(md, isSelf = false, muted = false, suppressed = false))
-        assertTrue(ChanACL.canOfferMute(ChanACL.WRITE, isSelf = false, muted = false, suppressed = false))
-        assertFalse(ChanACL.canOfferMute(md, isSelf = true, muted = false, suppressed = false))
-        assertTrue(ChanACL.canOfferMute(md, isSelf = true, muted = false, suppressed = true))
-        assertTrue(ChanACL.canOfferMute(md, isSelf = true, muted = true, suppressed = false))
-        assertTrue(ChanACL.canOfferMute(md, isSelf = true, muted = true, suppressed = true))
+        assertTrue(ChanACL.canOfferMute(md.toLong(), isSelf = false, muted = false, suppressed = false))
+        assertTrue(ChanACL.canOfferMute(ChanACL.WRITE.toLong(), isSelf = false, muted = false, suppressed = false))
+        assertFalse(ChanACL.canOfferMute(md.toLong(), isSelf = true, muted = false, suppressed = false))
+        assertTrue(ChanACL.canOfferMute(md.toLong(), isSelf = true, muted = false, suppressed = true))
+        assertTrue(ChanACL.canOfferMute(md.toLong(), isSelf = true, muted = true, suppressed = false))
+        assertTrue(ChanACL.canOfferMute(md.toLong(), isSelf = true, muted = true, suppressed = true))
     }
 
     @Test
     fun selfRegister_requiresSelfRegisterOrWrite() {
         assertFalse(ChanACL.canSelfRegister(0))
-        assertFalse(ChanACL.canSelfRegister(ChanACL.REGISTER))
-        assertTrue(ChanACL.canSelfRegister(ChanACL.SELF_REGISTER))
-        assertTrue(ChanACL.canSelfRegister(ChanACL.WRITE))
+        assertFalse(ChanACL.canSelfRegister(ChanACL.REGISTER.toLong()))
+        assertTrue(ChanACL.canSelfRegister(ChanACL.SELF_REGISTER.toLong()))
+        assertTrue(ChanACL.canSelfRegister(ChanACL.WRITE.toLong()))
     }
 
     @Test
     fun registerOthers_requiresRegisterOrWrite() {
         assertFalse(ChanACL.canRegisterOthers(0))
-        assertFalse(ChanACL.canRegisterOthers(ChanACL.SELF_REGISTER))
-        assertTrue(ChanACL.canRegisterOthers(ChanACL.REGISTER))
-        assertTrue(ChanACL.canRegisterOthers(ChanACL.WRITE))
+        assertFalse(ChanACL.canRegisterOthers(ChanACL.SELF_REGISTER.toLong()))
+        assertTrue(ChanACL.canRegisterOthers(ChanACL.REGISTER.toLong()))
+        assertTrue(ChanACL.canRegisterOthers(ChanACL.WRITE.toLong()))
     }
 
     @Test
@@ -86,32 +86,32 @@ class ChanACLTest {
         val adminReg = ChanACL.REGISTER
         assertTrue(
             ChanACL.canOfferRegister(
-                selfReg, isSelf = true, isRegistered = false, hasCertificate = true,
+                selfReg.toLong(), isSelf = true, isRegistered = false, hasCertificate = true,
             )
         )
         assertFalse(
             ChanACL.canOfferRegister(
-                selfReg, isSelf = false, isRegistered = false, hasCertificate = true,
+                selfReg.toLong(), isSelf = false, isRegistered = false, hasCertificate = true,
             )
         )
         assertTrue(
             ChanACL.canOfferRegister(
-                adminReg, isSelf = false, isRegistered = false, hasCertificate = true,
+                adminReg.toLong(), isSelf = false, isRegistered = false, hasCertificate = true,
             )
         )
         assertFalse(
             ChanACL.canOfferRegister(
-                adminReg, isSelf = true, isRegistered = false, hasCertificate = true,
+                adminReg.toLong(), isSelf = true, isRegistered = false, hasCertificate = true,
             )
         )
         assertFalse(
             ChanACL.canOfferRegister(
-                selfReg, isSelf = true, isRegistered = true, hasCertificate = true,
+                selfReg.toLong(), isSelf = true, isRegistered = true, hasCertificate = true,
             )
         )
         assertFalse(
             ChanACL.canOfferRegister(
-                selfReg, isSelf = true, isRegistered = false, hasCertificate = false,
+                selfReg.toLong(), isSelf = true, isRegistered = false, hasCertificate = false,
             )
         )
     }
@@ -120,29 +120,29 @@ class ChanACLTest {
     fun textMessage_requiresWriteOrTextMessage() {
         assertEquals(0x200, ChanACL.TEXT_MESSAGE)
         assertFalse(ChanACL.canTextMessage(0))
-        assertFalse(ChanACL.canTextMessage(ChanACL.ENTER))
-        assertTrue(ChanACL.canTextMessage(ChanACL.TEXT_MESSAGE))
-        assertTrue(ChanACL.canTextMessage(ChanACL.WRITE))
-        assertTrue(ChanACL.canTextMessage(ChanACL.WRITE or ChanACL.TEXT_MESSAGE))
+        assertFalse(ChanACL.canTextMessage(ChanACL.ENTER.toLong()))
+        assertTrue(ChanACL.canTextMessage(ChanACL.TEXT_MESSAGE.toLong()))
+        assertTrue(ChanACL.canTextMessage(ChanACL.WRITE.toLong()))
+        assertTrue(ChanACL.canTextMessage((ChanACL.WRITE or ChanACL.TEXT_MESSAGE).toLong()))
     }
 
     @Test
     fun listen_requiresWriteOrListen() {
         assertEquals(0x800, ChanACL.LISTEN)
         assertFalse(ChanACL.canListen(0))
-        assertFalse(ChanACL.canListen(ChanACL.TEXT_MESSAGE))
-        assertTrue(ChanACL.canListen(ChanACL.LISTEN))
-        assertTrue(ChanACL.canListen(ChanACL.WRITE))
-        assertTrue(ChanACL.canListen(ChanACL.WRITE or ChanACL.LISTEN))
+        assertFalse(ChanACL.canListen(ChanACL.TEXT_MESSAGE.toLong()))
+        assertTrue(ChanACL.canListen(ChanACL.LISTEN.toLong()))
+        assertTrue(ChanACL.canListen(ChanACL.WRITE.toLong()))
+        assertTrue(ChanACL.canListen((ChanACL.WRITE or ChanACL.LISTEN).toLong()))
     }
 
     @Test
     fun linkChannel_requiresWriteOrLinkChannel() {
         assertEquals(0x80, ChanACL.LINK_CHANNEL)
         assertFalse(ChanACL.canLinkChannel(0))
-        assertFalse(ChanACL.canLinkChannel(ChanACL.LISTEN))
-        assertTrue(ChanACL.canLinkChannel(ChanACL.LINK_CHANNEL))
-        assertTrue(ChanACL.canLinkChannel(ChanACL.WRITE))
+        assertFalse(ChanACL.canLinkChannel(ChanACL.LISTEN.toLong()))
+        assertTrue(ChanACL.canLinkChannel(ChanACL.LINK_CHANNEL.toLong()))
+        assertTrue(ChanACL.canLinkChannel(ChanACL.WRITE.toLong()))
     }
 
     @Test
@@ -150,15 +150,15 @@ class ChanACLTest {
         assertEquals(0x40, ChanACL.MAKE_CHANNEL)
         assertEquals(0x400, ChanACL.MAKE_TEMP_CHANNEL)
         assertFalse(ChanACL.canAddChannel(0))
-        assertFalse(ChanACL.canAddChannel(ChanACL.ENTER))
-        assertTrue(ChanACL.canAddChannel(ChanACL.MAKE_CHANNEL))
-        assertTrue(ChanACL.canAddChannel(ChanACL.MAKE_TEMP_CHANNEL))
-        assertTrue(ChanACL.canAddChannel(ChanACL.WRITE))
-        assertFalse(ChanACL.canMakePermanentChannel(ChanACL.MAKE_TEMP_CHANNEL))
-        assertTrue(ChanACL.canMakePermanentChannel(ChanACL.MAKE_CHANNEL))
-        assertTrue(ChanACL.canMakePermanentChannel(ChanACL.WRITE))
+        assertFalse(ChanACL.canAddChannel(ChanACL.ENTER.toLong()))
+        assertTrue(ChanACL.canAddChannel(ChanACL.MAKE_CHANNEL.toLong()))
+        assertTrue(ChanACL.canAddChannel(ChanACL.MAKE_TEMP_CHANNEL.toLong()))
+        assertTrue(ChanACL.canAddChannel(ChanACL.WRITE.toLong()))
+        assertFalse(ChanACL.canMakePermanentChannel(ChanACL.MAKE_TEMP_CHANNEL.toLong()))
+        assertTrue(ChanACL.canMakePermanentChannel(ChanACL.MAKE_CHANNEL.toLong()))
+        assertTrue(ChanACL.canMakePermanentChannel(ChanACL.WRITE.toLong()))
         assertFalse(ChanACL.canWrite(0))
-        assertTrue(ChanACL.canWrite(ChanACL.WRITE))
+        assertTrue(ChanACL.canWrite(ChanACL.WRITE.toLong()))
     }
 
     @Test
@@ -196,9 +196,9 @@ class ChanACLTest {
                 ChanACL.TEXT_MESSAGE or ChanACL.LISTEN,
             ChanACL.DEFAULT,
         )
-        assertFalse(ChanACL.has(ChanACL.SUPERUSER_EFFECTIVE, ChanACL.SPEAK))
-        assertFalse(ChanACL.has(ChanACL.SUPERUSER_EFFECTIVE, ChanACL.WHISPER))
-        assertTrue(ChanACL.has(ChanACL.SUPERUSER_EFFECTIVE, ChanACL.WRITE))
+        assertFalse(ChanACL.has(ChanACL.SUPERUSER_EFFECTIVE.toLong(), ChanACL.SPEAK))
+        assertFalse(ChanACL.has(ChanACL.SUPERUSER_EFFECTIVE.toLong(), ChanACL.WHISPER))
+        assertTrue(ChanACL.has(ChanACL.SUPERUSER_EFFECTIVE.toLong(), ChanACL.WRITE))
     }
 
     @Test
@@ -217,33 +217,33 @@ class ChanACLTest {
 
     @Test
     fun speakAndWhisper_areNotImpliedByWrite() {
-        assertFalse(ChanACL.canSpeak(ChanACL.WRITE))
-        assertFalse(ChanACL.canWhisper(ChanACL.WRITE))
-        assertTrue(ChanACL.canSpeak(ChanACL.SPEAK))
-        assertTrue(ChanACL.canWhisper(ChanACL.WHISPER))
-        assertTrue(ChanACL.canTraverse(ChanACL.WRITE))
-        assertTrue(ChanACL.canEnter(ChanACL.WRITE))
-        assertTrue(ChanACL.canResetUserContent(ChanACL.WRITE))
-        assertTrue(ChanACL.canResetUserContent(ChanACL.RESET_USER_CONTENT))
+        assertFalse(ChanACL.canSpeak(ChanACL.WRITE.toLong()))
+        assertFalse(ChanACL.canWhisper(ChanACL.WRITE.toLong()))
+        assertTrue(ChanACL.canSpeak(ChanACL.SPEAK.toLong()))
+        assertTrue(ChanACL.canWhisper(ChanACL.WHISPER.toLong()))
+        assertTrue(ChanACL.canTraverse(ChanACL.WRITE.toLong()))
+        assertTrue(ChanACL.canEnter(ChanACL.WRITE.toLong()))
+        assertTrue(ChanACL.canResetUserContent(ChanACL.WRITE.toLong()))
+        assertTrue(ChanACL.canResetUserContent(ChanACL.RESET_USER_CONTENT.toLong()))
     }
 
     @Test
     fun joinAndEditAcl_matchDesktopMenus() {
-        assertTrue(ChanACL.canJoinChannel(ChanACL.ENTER))
-        assertTrue(ChanACL.canJoinChannel(ChanACL.WRITE))
-        assertFalse(ChanACL.canJoinChannel(ChanACL.SPEAK))
-        assertTrue(ChanACL.canEditAcl(ChanACL.WRITE, 0))
-        assertTrue(ChanACL.canEditAcl(0, ChanACL.WRITE))
-        assertFalse(ChanACL.canEditAcl(ChanACL.ENTER, ChanACL.REGISTER))
+        assertTrue(ChanACL.canJoinChannel(ChanACL.ENTER.toLong()))
+        assertTrue(ChanACL.canJoinChannel(ChanACL.WRITE.toLong()))
+        assertFalse(ChanACL.canJoinChannel(ChanACL.SPEAK.toLong()))
+        assertTrue(ChanACL.canEditAcl(ChanACL.WRITE.toLong(), 0))
+        assertTrue(ChanACL.canEditAcl(0, ChanACL.WRITE.toLong()))
+        assertFalse(ChanACL.canEditAcl(ChanACL.ENTER.toLong(), ChanACL.REGISTER.toLong()))
     }
 
     @Test
     fun viewUserInfo_matchesDesktopMenu() {
         assertTrue(ChanACL.canViewUserInfo(0, 0, isSelf = true))
-        assertTrue(ChanACL.canViewUserInfo(ChanACL.REGISTER, 0, isSelf = false))
-        assertTrue(ChanACL.canViewUserInfo(ChanACL.WRITE, 0, isSelf = false))
-        assertTrue(ChanACL.canViewUserInfo(0, ChanACL.ENTER, isSelf = false))
-        assertFalse(ChanACL.canViewUserInfo(0, ChanACL.SPEAK, isSelf = false))
+        assertTrue(ChanACL.canViewUserInfo(ChanACL.REGISTER.toLong(), 0, isSelf = false))
+        assertTrue(ChanACL.canViewUserInfo(ChanACL.WRITE.toLong(), 0, isSelf = false))
+        assertTrue(ChanACL.canViewUserInfo(0, ChanACL.ENTER.toLong(), isSelf = false))
+        assertFalse(ChanACL.canViewUserInfo(0, ChanACL.SPEAK.toLong(), isSelf = false))
     }
 
     @Test
